[GH-5242] Moving wishlist to the proper dir
