+++ projects/storefrontlib/src/cms-components/wishlist/components/wish-list/wishlist.component.html
